fix: moves updateAddress back to recipientAddress
